requirements.txt
# app/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
langchain
langchain-google-genai
langgraph
aiofiles
python-dotenv
redis
pydantic-settings
pydub
boto3

# Testing libraries
pytest
pytest-asyncio
httpx

# Code Quality Tools
black
isort
flake8
mypy
pre-commit
bandit